+++ src/components/pages/History/ComputeJobs/Details.module.css
@@ -1,0 +1,2 @@
+.title {
+}
