+++ src/components/pages/History/index.tsx
@@ -1,11 +1,11 @@
 import React, { ReactElement } from 'react'
-import ComputeJobs from './ComputeJobs'
-import styles from './index.module.css'
+import Tabs from '../../atoms/Tabs'
 import PoolShares from './PoolShares'
 import PoolTransactions from '../../molecules/PoolTransactions'
 import PublishedList from './PublishedList'
 import Downloads from './Downloads'
-import Tabs from '../../atoms/Tabs'
+import ComputeJobs from './ComputeJobs'
+import styles from './index.module.css'
 
 const tabs = [
   {
